separate data preprocess to module
